+++ include/ontoloGenius/core/ontoGraphs/Graphs/DataPropertyGraph.h
@@ -49,7 +49,7 @@ private:
     if(find)
       return;
 
-    for(unsigned int i = 0; i < vect.size(); i++)
+    for(size_t i = 0; i < vect.size(); i++)
       if(disjoint == vect[i]->value_)
       {
         me->setSteady_disjoint(vect[i]);
@@ -65,7 +65,7 @@ private:
     if(find)
       return;
 
-    for(unsigned int i = 0; i < vect.size(); i++)
+    for(size_t i = 0; i < vect.size(); i++)
       if(domain == vect[i]->value_)
       {
         me->setSteady_domain(vect[i]);
